@@ -1,7 +1,7 @@
 
 [tool.poetry]
 name = "omagent_core"
-version = "0.2.2"
+version = "0.2.4"
 description = "Core package for OmAgent"
 authors = ["OM AI Lab <OmAIglobal2024@gmail.com>"]
 readme = "README.md"
@@ -13,7 +13,7 @@ classifiers = [
 repository = "https://github.com/om-ai-lab/OmAgent"
 
 [tool.poetry.dependencies]
-python = "^3.10,<3.12"
+python = "^3.11,<3.12"
 sqlmodel = "^0.0.22"
 pymysql = "^1.1.1"
 sqlalchemy-utils = "^0.41.2"
@@ -43,10 +43,16 @@ redis = "^5.2.1"
 json-repair = "^0.31.0"
 scenedetect = "^0.6.5.1"
 pydub = "^0.25.1"
-qwen_vl_utils = "^3.4.0"
+qwen_vl_utils = "0.0.8"
 face-recognition = "^1.3.0"
+tavily-python = "^0.5.0"
 httpx = {extras = ["http2"], version = "^0.28.1"}
 func-timeout = "^4.3.5"
+gradio = "^5.7.1"
+opencv-python = "^4.10.0.84"
+wikipedia = "^1.4.0"
+sympy = "^1.13.3"
+redislite = "^6.2.912183"
 
 [build-system]
 requires = ["poetry-core"]
