@@ -1,11 +1,12 @@
 # Import required modules and components
 import os
+os.environ["OMAGENT_MODE"] = "lite"
 from pathlib import Path
 
 from agent.conclude.conclude import Conclude
 from agent.input_interface.input_interface import InputInterface
 from omagent_core.advanced_components.workflow.dnc.workflow import DnCWorkflow
-from omagent_core.clients.devices.cli.client import DefaultClient
+from omagent_core.clients.devices.cli import DefaultClient
 from omagent_core.engine.workflow.conductor_workflow import ConductorWorkflow
 from omagent_core.engine.workflow.task.simple_task import simple_task
 from omagent_core.utils.container import container
