@@ -1,5 +1,4 @@
 # Import required modules and components
-import os
 from omagent_core.utils.container import container
 from omagent_core.engine.workflow.conductor_workflow import ConductorWorkflow
 from omagent_core.engine.workflow.task.simple_task import simple_task
@@ -8,7 +7,8 @@ from dotenv import load_dotenv
 load_dotenv()  
 
 from omagent_core.utils.registry import registry
-from omagent_core.clients.devices.cli.client import DefaultClient
+from omagent_core.clients.devices.cli import DefaultClient
+
 from omagent_core.utils.logger import logging
 from agent.input_interface.input_interface import COTInputInterface
 from omagent_core.advanced_components.workflow.self_consist_cot.workflow import SelfConsistentWorkflow
@@ -29,7 +29,7 @@ with open(config_path.joinpath('path_config.yaml'), 'r') as f:
 registry.import_module(CURRENT_PATH.joinpath('agent'))
 
 # Load container configuration from YAML file
-container.register_stm("RedisSTM")
+container.register_stm("SharedMemSTM")
 container.from_config(CURRENT_PATH.joinpath('container.yaml'))
 
 # Initialize general_self_consist_cot workflow
