@@ -12,7 +12,7 @@ from pathlib import Path
 CURRENT_PATH = Path(__file__).parents[0]
 
 # Register core workflow components for state management, callbacks and input handling
-container.register_stm(stm="RedisSTM")
+container.register_stm("SharedMemSTM")
 container.register_callback(callback="AppCallback")
 container.register_input(input="AppInput")
 
