@@ -11,7 +11,7 @@ CURRENT_PATH = Path(__file__).parents[0]
 
 
 # Register required components
-container.register_stm("RedisSTM")
+container.register_stm("SharedMemSTM")
 container.register_callback(callback="AppCallback")
 container.register_input(input="AppInput")
 
