@@ -7,8 +7,8 @@ CURRENT_PATH = Path(__file__).parents[0]
 
 registry.import_module()
 
-container.register_stm(stm="RedisSTM")
-container.register_ltm(ltm="MilvusLTM")
+container.register_stm("SharedMemSTM")
+container.register_ltm("MilvusLTM")
 container.register_callback(callback="AppCallback")
 container.register_input(input="AppInput")
 
