@@ -1,3 +1,5 @@
+import os
+os.environ["OMAGENT_MODE"] = "lite"
 from omagent_core.engine.workflow.conductor_workflow import ConductorWorkflow
 from omagent_core.engine.workflow.task.simple_task import simple_task
 from omagent_core.utils.container import container
@@ -7,7 +9,7 @@ logging.init_logger("omagent", "omagent", level="INFO")
 
 from pathlib import Path
 
-from omagent_core.clients.devices.webpage.client import WebpageClient
+from omagent_core.clients.devices.webpage import WebpageClient
 
 CURRENT_PATH = Path(__file__).parents[0]
 
@@ -15,7 +17,6 @@ from omagent_core.utils.registry import registry
 
 registry.import_module(project_path=CURRENT_PATH.joinpath("agent"))
 
-import os
 import sys
 
 sys.path.append(os.path.abspath(CURRENT_PATH.joinpath("../../../")))
