@@ -45,14 +45,9 @@ class ReactProWorkflow(ConductorWorkflow):
         # Do-While loop with max_turns from config
         self.loop_task = DoWhileTask(
             task_ref_name='react_pro_loop',
-            tasks=[self.think_task, self.action_task, self.wiki_search_task],
-            termination_condition=f'''
-                if (($.action.is_final == true) || ($.think.step_number > $.think.max_steps)) {{
-                    false;  // Stop loop if it's a Finish action or exceeded max turns
-                }} else {{
-                    true;   // Continue loop otherwise
-                }}
-            '''
+            tasks=[self.think_task, self.action_task, self.wiki_search_task], #'$.action[is_final] == true' 
+            #(($.action.is_final == true) || ($.think.step_number > $.think.max_steps))
+            termination_condition='(($.action.is_final == true) || ($.think.step_number > $.think.max_steps))' 
         )
         
         # Configure output task
@@ -60,8 +55,7 @@ class ReactProWorkflow(ConductorWorkflow):
             task_def_name=ReactOutput,
             task_reference_name='react_output',
             inputs={
-                'action_output': '${action.output}',
-                'workflow_id': '${workflow.workflowId}'
+                'action_output': '${action.output}'
             }
         )
         
