@@ -10,6 +10,8 @@ from omagent_core.models.llms.prompt.prompt import PromptTemplate
 from pydantic import Field
 from typing import List
 
+from pydantic_core.core_schema import NoInfoWrapValidatorFunctionSchema
+
 CURRENT_PATH = Path(__file__).parents[0]
 
 @registry.register_worker()
@@ -47,54 +49,22 @@ class COTReasoning(BaseLLMBackend, BaseWorker):
         Returns:
             dict: A dictionary containing the reasoning results.
         """
-
-        reason_path = []
-        prompt_token = []
-        complete_token = []
-        for i in range(num_path):
-            reasoning_result = self.simple_infer(question=user_question, examples=examples)
-            prompt_token.append(reasoning_result["usage"]["prompt_tokens"])
-            complete_token.append(reasoning_result["usage"]["completion_tokens"])
-            reasoning_result = reasoning_result["choices"][0]["message"]["content"]
-            reason_path.append(reasoning_result)
-            if i == 1:
-                message = self.prep_prompt([{"question": user_question}])
-                body = self.llm._msg2req(message[0])
-                self.stm(self.workflow_instance_id)["body"] = body
-
-        self.stm(self.workflow_instance_id)['prompt_token'] = prompt_token
-        self.stm(self.workflow_instance_id)['completion_token'] = complete_token
+        if self.llm.model_id in ["gpt4o","gpt4o-mini","gpt3.5-turbo"]:
+            reasoning_result = self.infer(input_list=[{"question": user_question, "examples": examples}], n=num_path)[0]
+            reason_path = [choice["message"]["content"] for choice in reasoning_result["choices"]]
+            message = self.prep_prompt([{"question": user_question, "examples": examples}])
+            body = self.llm._msg2req(message[0])
+            self.stm(self.workflow_instance_id)["body"] = body
+        else:
+            reason_path = []
+            for i in range(num_path):
+                reasoning_result = self.simple_infer(question=user_question,examples=examples)
+                reasoning_result = reasoning_result["choices"][0]["message"]["content"]
+                reason_path.append(reasoning_result)
+                if i==1:
+                    message = self.prep_prompt([{"question":user_question, "examples": examples}])
+                    body = self.llm._msg2req(message[0])
+                    self.stm(self.workflow_instance_id)["body"] = body
         self.stm(self.workflow_instance_id)['reasoning_result'] = reason_path
-
         self.callback.send_answer(self.workflow_instance_id, msg=",".join(reason_path))
-
-        return {'reasoning_result': reason_path}
-
-
-    def _run(self, user_question: str, num_path: int, examples: str = None, *args, **kwargs):
-        reason_path = []
-        prompt_token = []
-        complete_token = []
-        for i in range(num_path):
-            reasoning_result = self.simple_infer(question=user_question,examples=examples)
-            prompt_token.append(reasoning_result["usage"]["prompt_tokens"])
-            complete_token.append(reasoning_result["usage"]["completion_tokens"])
-            reasoning_result = reasoning_result["choices"][0]["message"]["content"]
-            reason_path.append(reasoning_result)
-            if i==1:
-                message = self.prep_prompt([{"question":user_question}])
-                body = self.llm._msg2req(message[0])
-                self.stm(self.workflow_instance_id)["body"] = body
-
-
-
-        self.stm(self.workflow_instance_id)['prompt_token'] = prompt_token
-        self.stm(self.workflow_instance_id)['completion_token'] = complete_token
-        self.stm(self.workflow_instance_id)['reasoning_result'] = reason_path
-
-        self.callback.send_answer(self.workflow_instance_id, msg=",".join(reason_path))
-
-
-
-
         return {'reasoning_result': reason_path}
