@@ -42,6 +42,7 @@ class DefaultClient:
         self._config_path = config_path
         self._workers = workers
         self._input_prompt = input_prompt
+        self._task_to_domain = {}
 
     def start_interactor(self):
         workflow_instance_id = None
@@ -49,11 +50,11 @@ class DefaultClient:
             absolute_path = Path(self._config_path).resolve()
             worker_config = build_from_file(self._config_path)
             self._task_handler_interactor = TaskHandler(
-                worker_config=worker_config, workers=self._workers
+                worker_config=worker_config, workers=self._workers, task_to_domain=self._task_to_domain
             )
             self._task_handler_interactor.start_processes()
             workflow_instance_id = self._interactor.start_workflow_with_input(
-                workflow_input={}
+                workflow_input={}, task_to_domain=self._task_to_domain
             )
 
             stream_name = f"{workflow_instance_id}_output"
@@ -180,11 +181,11 @@ class DefaultClient:
         try:
             worker_config = build_from_file(self._config_path)
             self._task_handler_processor = TaskHandler(
-                worker_config=worker_config, workers=self._workers
+                worker_config=worker_config, workers=self._workers, task_to_domain=self._task_to_domain
             )
             self._task_handler_processor.start_processes()
             workflow_instance_id = self._processor.start_workflow_with_input(
-                workflow_input={}
+                workflow_input={}, task_to_domain=self._task_to_domain
             )
             user_input = input(
                 f"{Fore.GREEN}Please input a folder path of images:(WaitPress Enter to finish the entire input.):\n>>>{Style.RESET_ALL}"
