@@ -21,14 +21,15 @@ from omagent_core.utils.logger import logging
 from omagent_core.utils.registry import registry
 from shortuuid import uuid
 from typing_extensions import Self
-
-import json
+import os
+from omagent_core.engine.http.models.workflow_def import to_workflow_def
+from omagent_core.engine.workflow.task.simple_task import simple_task
 
 class ConductorWorkflow:
     SCHEMA_VERSION = 2
 
     def __init__(self, name: str, version: int = None, description: str = None, lite_version: bool = False) -> Self:
-        if lite_version:
+        if lite_version or os.getenv("OMAGENT_MODE") == "lite":
             self._executor = LiteWorkflowExecutor()
         else:
             self._executor = WorkflowExecutor()
@@ -194,6 +195,31 @@ class ConductorWorkflow:
         keys = list(input.keys())
         self.input_template(input)
         return self
+        
+    def load(self, json_file_path: str) -> None:
+        import json
+        from pathlib import Path
+
+        # Load the JSON file
+        json_path = Path(json_file_path)
+        if not json_path.is_file():
+            raise FileNotFoundError(f"The file {json_file_path} does not exist.")
+
+        with open(json_path, 'r') as file:
+            data = json.load(file)
+        workflow_def = to_workflow_def(json_data=data)
+        self.name = workflow_def.name
+        self._tasks = [simple_task(task_def_name=task.name, task_reference_name=task.task_reference_name, inputs=task.input_parameters) for task in workflow_def.tasks]
+        self._input_parameters = workflow_def.input_parameters
+        self._output_parameters = workflow_def.output_parameters
+        self._failure_workflow = workflow_def.failure_workflow
+        self._timeout_seconds = workflow_def.timeout_seconds
+        self._variables = workflow_def.variables
+        self._input_template = workflow_def.input_template
+        self._workflow_status_listener_enabled = workflow_def.workflow_status_listener_enabled
+        self._owner_email = workflow_def.owner_email
+
+    
 
     # Register the workflow definition with the server. If overwrite is set, the definition on the server will be
     # overwritten. When not set, the call fails if there is any change in the workflow definition between the server
@@ -321,7 +347,9 @@ class ConductorWorkflow:
     def __get_workflow_task_list(self) -> List[WorkflowTask]:
         workflow_task_list = []
         for task in self._tasks:
+            # print (type(task))
             converted_task = task.to_workflow_task()
+            # print (converted_task)
             if isinstance(converted_task, list):
                 for subtask in converted_task:
                     workflow_task_list.append(subtask)
