@@ -2,7 +2,6 @@ import os
 from datetime import datetime
 from typing import Any, Dict, List
 
-from transformers import AutoTokenizer, AutoProcessor, Qwen2VLForConditionalGeneration
 from pydantic import Field
 from .schemas import Content, Message
 from ...utils.registry import registry
@@ -31,6 +30,7 @@ class Qwen2_VL(BaseLLM):
 
     def __init__(self, **data: Any) -> None:
         super().__init__(**data)
+        from transformers import AutoTokenizer, AutoProcessor, Qwen2VLForConditionalGeneration
         self.processor = AutoProcessor.from_pretrained(self.model_name,min_pixels=256 * 28 * 28, max_pixels=512 * 28 * 28)
         self.model = Qwen2VLForConditionalGeneration.from_pretrained(self.model_name, torch_dtype="auto", device_map="auto").to(self.device)
         
